+++ src/Network/Network/Common/NetworkResourceManagerProfile.cs
@@ -578,6 +578,10 @@ namespace Microsoft.Azure.Commands.Network
                     .ForMember(
                         dest => dest.IPTagType,
                         opt => opt.MapFrom(src => src.IpTagType)
+                    )
+                    .ForMember(
+                        dest => dest.FirstPartyServiceTagId,
+                        opt => opt.MapFrom(src => src.FirstPartyServiceTagId)
                     );
                 cfg.CreateMap<CNM.PSPublicIpAddress, MNM.ReferencedPublicIpAddress>();
 
@@ -592,6 +596,10 @@ namespace Microsoft.Azure.Commands.Network
                     .ForMember(
                         dest => dest.IpTagType,
                         opt => opt.MapFrom(src => src.IPTagType)
+                    )
+                    .ForMember(
+                        dest => dest.FirstPartyServiceTagId,
+                        opt => opt.MapFrom(src => src.FirstPartyServiceTagId)
                     );
                 cfg.CreateMap<MNM.ReferencedPublicIpAddress, CNM.PSPublicIpAddress>();
 
@@ -1171,6 +1179,10 @@ namespace Microsoft.Azure.Commands.Network
                 cfg.CreateMap<MNM.ExpressRouteLink, CNM.PSExpressRouteLink>();
                 cfg.CreateMap<MNM.ExpressRoutePortAuthorization, CNM.PSExpressRoutePortAuthorization>();
 
+                // ExpressRouteAuthorizationKey (shared by circuit and port authorization listKeys)
+                cfg.CreateMap<CNM.PSExpressRouteAuthorizationKey, MNM.ExpressRouteAuthorizationKey>();
+                cfg.CreateMap<MNM.ExpressRouteAuthorizationKey, CNM.PSExpressRouteAuthorizationKey>();
+
                 // ExpressRouteLag
                 // The SDK model is nested (Properties of type ExpressRouteLagPropertiesFormat) while the
                 // PowerShell model is flat, so bridge the flat PS members to/from the nested SDK Properties.
@@ -1298,6 +1310,15 @@ namespace Microsoft.Azure.Commands.Network
                 // ExpressRouteCrossConnection
                 // CNM to MNM
                 cfg.CreateMap<CNM.PSExpressRouteCrossConnection, MNM.ExpressRouteCrossConnection>();
+                cfg.CreateMap<CNM.PSExpressRouteCrossConnectionPortMapping, MNM.PortMapping>();
+                cfg.CreateMap<MNM.MigrateExpressRouteCircuitValidateResponse, CNM.PSExpressRouteCircuitMigrationValidationResult>();
+                cfg.CreateMap<MNM.MigrateExpressRouteCircuitHealthCheckResponse, CNM.PSExpressRouteCircuitMigrationResult>();
+                cfg.CreateMap<MNM.MigrateExpressRouteCircuitHealthCheckDetails, CNM.PSExpressRouteCircuitMigrationHealthDetails>();
+                cfg.CreateMap<MNM.PortMigrationInfo, CNM.PSExpressRouteCircuitMigrationPortInfo>();
+                cfg.CreateMap<MNM.PeeringHealth, CNM.PSExpressRouteCircuitMigrationPeeringHealth>();
+                cfg.CreateMap<MNM.PeeringStats, CNM.PSExpressRouteCircuitMigrationPeeringStats>();
+                cfg.CreateMap<MNM.SourcePortStats, CNM.PSExpressRouteCircuitMigrationSourcePortStats>();
+                cfg.CreateMap<MNM.Metric, CNM.PSExpressRouteCircuitMigrationMetric>();
                 cfg.CreateMap<CNM.PSExpressRouteCircuitReference, MNM.ExpressRouteCircuitReference>();
                 cfg.CreateMap<CNM.PSExpressRouteCrossConnectionPeering, MNM.ExpressRouteCrossConnectionPeering>()
                     .ForMember(
@@ -1548,6 +1569,18 @@ namespace Microsoft.Azure.Commands.Network
                     );
                 cfg.CreateMap<CNM.PSApplicationGatewayPathRule, MNM.ApplicationGatewayPathRule>();
                 cfg.CreateMap<CNM.PSApplicationGatewayUrlPathMap, MNM.ApplicationGatewayUrlPathMap>();
+                cfg.CreateMap<CNM.PSApplicationGatewayAdvancedRoutingPropertyValueMatcher, MNM.ApplicationGatewayAdvancedRoutingPropertyValueMatcher>();
+                cfg.CreateMap<CNM.PSApplicationGatewayAdvancedRoutingCondition, MNM.ApplicationGatewayAdvancedRoutingCondition>();
+                // The advanced routing SDK models nest their payload under Properties instead of flattening it.
+                cfg.CreateMap<CNM.PSApplicationGatewayAdvancedRoutingConditionSet, MNM.ApplicationGatewayAdvancedRoutingConditionSetPropertiesFormat>();
+                cfg.CreateMap<CNM.PSApplicationGatewayAdvancedRoutingConditionSet, MNM.ApplicationGatewayAdvancedRoutingConditionSet>()
+                    .ForMember(dest => dest.Properties, opt => opt.MapFrom(src => src));
+                cfg.CreateMap<CNM.PSApplicationGatewayAdvancedRoutingRule, MNM.ApplicationGatewayAdvancedRoutingRulePropertiesFormat>();
+                cfg.CreateMap<CNM.PSApplicationGatewayAdvancedRoutingRule, MNM.ApplicationGatewayAdvancedRoutingRule>()
+                    .ForMember(dest => dest.Properties, opt => opt.MapFrom(src => src));
+                cfg.CreateMap<CNM.PSApplicationGatewayAdvancedRoutingMap, MNM.ApplicationGatewayAdvancedRoutingMapPropertiesFormat>();
+                cfg.CreateMap<CNM.PSApplicationGatewayAdvancedRoutingMap, MNM.ApplicationGatewayAdvancedRoutingMap>()
+                    .ForMember(dest => dest.Properties, opt => opt.MapFrom(src => src));
                 cfg.CreateMap<CNM.PSApplicationGatewayProbeHealthResponseMatch, MNM.ApplicationGatewayProbeHealthResponseMatch>()
                     .AfterMap((src, dest) => dest.StatusCodes = (src.StatusCodes == null) ? null : dest.StatusCodes);
                 cfg.CreateMap<CNM.PSApplicationGatewayProbe, MNM.ApplicationGatewayProbe>();
@@ -1658,6 +1691,26 @@ namespace Microsoft.Azure.Commands.Network
                     );
                 cfg.CreateMap<MNM.ApplicationGatewayPathRule, CNM.PSApplicationGatewayPathRule>();
                 cfg.CreateMap<MNM.ApplicationGatewayUrlPathMap, CNM.PSApplicationGatewayUrlPathMap>();
+                cfg.CreateMap<MNM.ApplicationGatewayAdvancedRoutingPropertyValueMatcher, CNM.PSApplicationGatewayAdvancedRoutingPropertyValueMatcher>();
+                cfg.CreateMap<MNM.ApplicationGatewayAdvancedRoutingCondition, CNM.PSApplicationGatewayAdvancedRoutingCondition>();
+                cfg.CreateMap<MNM.ApplicationGatewayAdvancedRoutingConditionSet, CNM.PSApplicationGatewayAdvancedRoutingConditionSet>()
+                    .ForMember(dest => dest.RoutingConditions, opt => opt.MapFrom(src => src.Properties.RoutingConditions))
+                    .ForMember(dest => dest.ProvisioningState, opt => opt.MapFrom(src => src.Properties.ProvisioningState));
+                cfg.CreateMap<MNM.ApplicationGatewayAdvancedRoutingRule, CNM.PSApplicationGatewayAdvancedRoutingRule>()
+                    .ForMember(dest => dest.Priority, opt => opt.MapFrom(src => src.Properties.Priority))
+                    .ForMember(dest => dest.AdvancedRoutingConditionSet, opt => opt.MapFrom(src => src.Properties.AdvancedRoutingConditionSet))
+                    .ForMember(dest => dest.BackendAddressPool, opt => opt.MapFrom(src => src.Properties.BackendAddressPool))
+                    .ForMember(dest => dest.BackendHttpSettings, opt => opt.MapFrom(src => src.Properties.BackendHttpSettings))
+                    .ForMember(dest => dest.RedirectConfiguration, opt => opt.MapFrom(src => src.Properties.RedirectConfiguration))
+                    .ForMember(dest => dest.RewriteRuleSet, opt => opt.MapFrom(src => src.Properties.RewriteRuleSet))
+                    .ForMember(dest => dest.ProvisioningState, opt => opt.MapFrom(src => src.Properties.ProvisioningState));
+                cfg.CreateMap<MNM.ApplicationGatewayAdvancedRoutingMap, CNM.PSApplicationGatewayAdvancedRoutingMap>()
+                    .ForMember(dest => dest.DefaultBackendAddressPool, opt => opt.MapFrom(src => src.Properties.DefaultBackendAddressPool))
+                    .ForMember(dest => dest.DefaultBackendHttpSettings, opt => opt.MapFrom(src => src.Properties.DefaultBackendHttpSettings))
+                    .ForMember(dest => dest.DefaultRedirectConfiguration, opt => opt.MapFrom(src => src.Properties.DefaultRedirectConfiguration))
+                    .ForMember(dest => dest.DefaultRewriteRuleSet, opt => opt.MapFrom(src => src.Properties.DefaultRewriteRuleSet))
+                    .ForMember(dest => dest.AdvancedRoutingRules, opt => opt.MapFrom(src => src.Properties.AdvancedRoutingRules))
+                    .ForMember(dest => dest.ProvisioningState, opt => opt.MapFrom(src => src.Properties.ProvisioningState));
                 cfg.CreateMap<MNM.ApplicationGatewayProbeHealthResponseMatch, CNM.PSApplicationGatewayProbeHealthResponseMatch>()
                     .AfterMap((src, dest) => dest.StatusCodes = (src.StatusCodes == null) ? null : dest.StatusCodes);
                 cfg.CreateMap<MNM.ApplicationGatewayProbe, CNM.PSApplicationGatewayProbe>();