+++ src/Network/Network/Models/PSExpressRouteCircuitMigrationHealthDetails.cs
@@ -1,0 +1,9 @@
+using System.Collections.Generic;
+
+namespace Microsoft.Azure.Commands.Network.Models
+{
+    public class PSExpressRouteCircuitMigrationHealthDetails
+    {
+        public List<PSExpressRouteCircuitMigrationPortInfo> PortMigrationInfos { get; set; }
+    }
+}